# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This service automatically sends messages and provides control over the sending process",
        "title": "Message Sending System API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:4300",
    "basePath": "/",
    "paths": {
        "/messages": {
            "get": {
                "description": "Retrieve a list of all sent messages from the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "messages"
                ],
                "summary": "List sent messages",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.MessageListResponse"
                        }
                    },
                    "404": {
                        "description": "No messages found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to fetch messages",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/worker": {
            "post": {
                "description": "Start or stop the automatic message sending worker",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "worker"
                ],
                "summary": "Control message sending worker",
                "responses": {
                    "200": {
                        "description": "Worker started/stopped successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "405": {
                        "description": "Method not allowed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.MessageListResponse": {
            "description": "List of messages response structure",
            "type": "object",
            "properties": {
                "messages": {
                    "description": "Array of messages",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.MessageResponse"
                    }
                }
            }
        },
        "api.MessageResponse": {
            "description": "Single message response structure",
            "type": "object",
            "properties": {
                "content": {
                    "description": "The content of the message",
                    "type": "string",
                    "example": "Hello, this is a test message"
                },
                "id": {
                    "description": "The unique identifier of the message, big serial in database",
                    "type": "integer",
                    "example": 1
                },
                "recipient_phone_number": {
                    "description": "The recipient's phone number",
                    "type": "string",
                    "example": "+1234567890"
                },
                "status": {
                    "description": "The current status of the message (pending, sent, failed)",
                    "type": "string",
                    "example": "sent"
                }
            }
        }
    }
}